Rails3.1 migration: make cucumber run again
The first cucumber test passes now, all the others remain to be
ported. Also as visible in the many commented out lines, there's
still a lot to clean up.

This represents the technical spike only and a proof that we can
actually get back to running cucumber tests after the Rails3.1
migration.

It should help whomever will want to push further here.